+++ lib/services/offline/hive/adapters/hive_item.dart
@@ -1,4 +1,3 @@
-import 'dart:io';
 
 import 'package:hive/hive.dart';
 import 'package:impulse/app/app.dart';
