+++ test/widget_test.dart
@@ -12,7 +12,7 @@ import 'package:impulse/main.dart';
 void main() {
   testWidgets('Counter increments smoke test', (WidgetTester tester) async {
     // Build our app and trigger a frame.
-    await tester.pumpWidget(const Impulse());
+    await tester.pumpWidget(Impulse());
 
     // Verify that our counter starts at 0.
     expect(find.text('0'), findsOneWidget);
